refactor: moved Adapter classes to "adapters" package
