+++ app/src/main/java/campus/tech/kakao/map/views/adapters/SearchResultAdapter.kt
@@ -1,8 +1,7 @@
-package campus.tech.kakao.map.views
+package campus.tech.kakao.map.views.adapters
 
 import android.view.LayoutInflater
 import android.view.View
-import android.view.View.OnClickListener
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
